clk freq return uint32 is long enough

+++ esp_system/driver/include/hw/clk-tree.h
@@ -26,31 +26,31 @@ extern __attribute__((nothrow))
 extern __attribute__((nothrow))
     int CLK_pll_conf(enum PLL_freq_sel_t sel);
 extern __attribute__((nothrow, const))
-    uint64_t CLK_pll_freq(void);
+    uint32_t CLK_pll_freq(void);
 
     // CPU
 extern __attribute__((nothrow))
     int CLK_cpu_conf(enum CPU_sclk_sel_t sel, uint32_t div);
 extern __attribute__((nothrow, const))
-    uint64_t CLK_cpu_freq(void);
+    uint32_t CLK_cpu_freq(void);
 
     // SysTimer
 extern __attribute__((nothrow))
     int CLK_systimer_conf(enum SYSTIMER_sclk_sel_t sel);
 extern __attribute__((nothrow, const))
-    uint64_t CLK_systimer_freq(void);
+    uint32_t CLK_systimer_freq(void);
 
     // AHB/APB
 extern __attribute__((nothrow, const))
-    uint64_t CLK_ahb_freq(void);
+    uint32_t CLK_ahb_freq(void);
 extern __attribute__((nothrow, const))
-    uint64_t CLK_apb_freq(void);
+    uint32_t CLK_apb_freq(void);
 
     // RTC
 extern __attribute__((nothrow))
     int CLK_rtc_conf(enum RTC_sclk_sel_t sel);
 extern __attribute__((nothrow, const))
-    uint64_t CLK_rtc_freq(void);
+    uint32_t CLK_rtc_freq(void);
 
 /****************************************************************************
  *  peripheral module gating control
